move `getSelectedInternalAccount` from `selectors.js` to `accounts.ts`
