+++ test/data/mock-state.json
@@ -420,6 +420,7 @@
           "address": "0x0dcd5d886577d5081b0c52e242ef29e70be3e7bc",
           "id": "cf8dace4-9439-4bd4-b3a8-88c821c8fcb3",
           "metadata": {
+            "importTime": 0,
             "name": "Test Account",
             "keyring": {
               "type": "HD Key Tree"
@@ -439,6 +440,7 @@
           "address": "0xec1adf982415d2ef5ec55899b9bfb8bc0f29251b",
           "id": "07c2cfec-36c9-46c4-8115-3836d3ac9047",
           "metadata": {
+            "importTime": 0,
             "name": "Test Account 2",
             "keyring": {
               "type": "HD Key Tree"
@@ -458,6 +460,7 @@
           "address": "0xc42edfcc21ed14dda456aa0756c153f7985d8813",
           "id": "15e69915-2a1a-4019-93b3-916e11fd432f",
           "metadata": {
+            "importTime": 0,
             "name": "Ledger Hardware 2",
             "keyring": {
               "type": "Ledger Hardware"
@@ -477,6 +480,7 @@
           "address": "0xeb9e64b93097bc15f01f13eae97015c57ab64823",
           "id": "784225f4-d30b-4e77-a900-c8bbce735b88",
           "metadata": {
+            "importTime": 0,
             "name": "Test Account 3",
             "keyring": {
               "type": "HD Key Tree"
@@ -496,6 +500,7 @@
           "address": "0xca8f1F0245530118D0cf14a06b01Daf8f76Cf281",
           "id": "694225f4-d30b-4e77-a900-c8bbce735b42",
           "metadata": {
+            "importTime": 0,
             "name": "Test Account 4",
             "keyring": {
               "type": "Custody test"
@@ -515,11 +520,13 @@
           "address": "0xb552685e3d2790efd64a175b00d51f02cdafee5d",
           "id": "c3deeb99-ba0d-4a4e-a0aa-033fc1f79ae3",
           "metadata": {
+            "importTime": 0,
             "name": "Snap Account 1",
             "keyring": {
               "type": "Snap Keyring"
             },
             "snap": {
+              "enabled": true,
               "id": "snap-id",
               "name": "snap-name"
             }
