+++ ui/hooks/useMultichainSelector.ts
@@ -11,6 +11,7 @@ export function useMultichainSelector<
 ) {
   return useSelector((state: TState) => {
     // We either pass an account or fallback to the currently selected one
+    // @ts-expect-error state types don't match
     return selector(state, account || getSelectedInternalAccount(state));
   });
 }
