+++ ui/selectors/index.js
@@ -7,3 +7,4 @@ export * from './permissions';
 export * from './selectors';
 export * from './transactions';
 export * from './approvals';
+export * from './accounts';
