+++ ui/selectors/institutional/selectors.ts
@@ -1,5 +1,6 @@
 import { toChecksumAddress } from 'ethereumjs-util';
-import { getAccountType, getSelectedInternalAccount } from '../selectors';
+import { getAccountType } from '../selectors';
+import { getSelectedInternalAccount } from '../accounts';
 import { getProviderConfig } from '../../ducks/metamask/metamask';
 import { hexToDecimal } from '../../../shared/modules/conversion.utils';
 // TODO: Remove restricted import
@@ -166,6 +167,7 @@ export function getCustodianIconForAddress(state: State, address: string) {
 
 export function getIsCustodianSupportedChain(state: State) {
   try {
+    // @ts-expect-error state types don't match
     const selectedAccount = getSelectedInternalAccount(state);
     const accountType = getAccountType(state);
     const providerConfig = getProviderConfig(state);
@@ -207,6 +209,7 @@ export function getIsCustodianSupportedChain(state: State) {
 
 export function getMMIAddressFromModalOrAddress(state: State) {
   const modalAddress = state?.appState?.modal?.modalState?.props?.address;
+  // @ts-expect-error state types don't match
   const selectedAddress = getSelectedInternalAccount(state)?.address;
 
   return modalAddress || selectedAddress;
